+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/CassandraMailboxSessionMapperFactory.java
@@ -27,6 +27,7 @@ import org.apache.james.mailbox.cassandra.mail.CassandraAnnotationMapper;
 import org.apache.james.mailbox.cassandra.mail.CassandraAttachmentMapper;
 import org.apache.james.mailbox.cassandra.mail.CassandraImapUidDAO;
 import org.apache.james.mailbox.cassandra.mail.CassandraMailboxMapper;
+import org.apache.james.mailbox.cassandra.mail.CassandraMessageDAO;
 import org.apache.james.mailbox.cassandra.mail.CassandraMessageIdDAO;
 import org.apache.james.mailbox.cassandra.mail.CassandraMessageMapper;
 import org.apache.james.mailbox.cassandra.user.CassandraSubscriptionMapper;
@@ -54,17 +55,19 @@ public class CassandraMailboxSessionMapperFactory extends MailboxSessionMapperFa
     private final ModSeqProvider modSeqProvider;
     private final MessageIdProvider messageIdProvider;
     private final CassandraTypesProvider typesProvider;
+    private final CassandraMessageDAO messageDAO;
     private final CassandraMessageIdDAO messageIdDAO;
     private final CassandraImapUidDAO imapUidDAO;
     private int maxRetry;
 
     @Inject
     public CassandraMailboxSessionMapperFactory(UidProvider uidProvider, ModSeqProvider modSeqProvider, MessageIdProvider messageIdProvider, Session session, CassandraTypesProvider typesProvider,
-            CassandraMessageIdDAO messageIdDAO, CassandraImapUidDAO imapUidDAO) {
+            CassandraMessageDAO messageDAO, CassandraMessageIdDAO messageIdDAO, CassandraImapUidDAO imapUidDAO) {
         this.uidProvider = uidProvider;
         this.modSeqProvider = modSeqProvider;
         this.messageIdProvider = messageIdProvider;
         this.session = session;
+        this.messageDAO = messageDAO;
         this.messageIdDAO = messageIdDAO;
         this.imapUidDAO = imapUidDAO;
         this.maxRetry = DEFAULT_MAX_RETRY;
@@ -77,8 +80,8 @@ public class CassandraMailboxSessionMapperFactory extends MailboxSessionMapperFa
 
     @Override
     public CassandraMessageMapper createMessageMapper(MailboxSession mailboxSession) {
-        return new CassandraMessageMapper(session, uidProvider, modSeqProvider, messageIdProvider, null, maxRetry, typesProvider, createAttachmentMapper(mailboxSession),
-                messageIdDAO, imapUidDAO);
+        return new CassandraMessageMapper(session, uidProvider, modSeqProvider, messageIdProvider, null, maxRetry, createAttachmentMapper(mailboxSession),
+                messageDAO, messageIdDAO, imapUidDAO);
     }
 
     @Override
