+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/CassandraMessageMapper.java
@@ -21,56 +21,35 @@ package org.apache.james.mailbox.cassandra.mail;
 
 import static com.datastax.driver.core.querybuilder.QueryBuilder.decr;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.eq;
-import static com.datastax.driver.core.querybuilder.QueryBuilder.in;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.incr;
-import static com.datastax.driver.core.querybuilder.QueryBuilder.insertInto;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.select;
-import static com.datastax.driver.core.querybuilder.QueryBuilder.set;
 import static com.datastax.driver.core.querybuilder.QueryBuilder.update;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageIds.MESSAGE_ID;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.ATTACHMENTS;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.BODY;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.BODY_CONTENT;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.BODY_OCTECTS;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.BODY_START_OCTET;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.FIELDS;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.FULL_CONTENT_OCTETS;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.HEADERS;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.HEADER_CONTENT;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.INTERNAL_DATE;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.METADATA;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.MOD_SEQ;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.PROPERTIES;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.TABLE_NAME;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.TEXTUAL_LINE_COUNT;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.ANSWERED;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.DELETED;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.DRAFT;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.FLAGGED;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.RECENT;
 import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.SEEN;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.USER;
-import static org.apache.james.mailbox.cassandra.table.CassandraMessageTable.Flag.USER_FLAGS;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.ByteBuffer;
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import javax.mail.Flags;
 import javax.mail.Flags.Flag;
 import javax.mail.util.SharedByteArrayInputStream;
 
-import org.apache.james.backends.cassandra.init.CassandraTypesProvider;
-import org.apache.james.backends.cassandra.utils.CassandraConstants;
 import org.apache.james.backends.cassandra.utils.CassandraUtils;
 import org.apache.james.backends.cassandra.utils.FunctionRunnerWithRetry;
 import org.apache.james.mailbox.FlagsBuilder;
@@ -107,16 +86,12 @@ import org.apache.james.mailbox.store.mail.model.impl.SimpleProperty;
 import com.datastax.driver.core.ResultSet;
 import com.datastax.driver.core.Row;
 import com.datastax.driver.core.Session;
-import com.datastax.driver.core.Statement;
 import com.datastax.driver.core.UDTValue;
 import com.datastax.driver.core.querybuilder.Assignment;
-import com.datastax.driver.core.querybuilder.Select;
-import com.datastax.driver.core.querybuilder.Select.Where;
 import com.github.fge.lambdas.Throwing;
 import com.github.steveash.guavate.Guavate;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
-import com.google.common.io.ByteStreams;
 import com.google.common.primitives.Bytes;
 
 public class CassandraMessageMapper implements MessageMapper {
@@ -126,22 +101,22 @@ public class CassandraMessageMapper implements MessageMapper {
     private final MessageIdProvider messageIdProvider;
     private final MailboxSession mailboxSession;
     private final UidProvider uidProvider;
-    private final CassandraTypesProvider typesProvider;
     private final int maxRetries;
     private final AttachmentMapper attachmentMapper;
+    private final CassandraMessageDAO messageDAO;
     private final CassandraMessageIdDAO messageIdDAO;
     private final CassandraImapUidDAO imapUidDAO;
 
-    public CassandraMessageMapper(Session session, UidProvider uidProvider, ModSeqProvider modSeqProvider, MessageIdProvider messageIdProvider, MailboxSession mailboxSession, int maxRetries, CassandraTypesProvider typesProvider, AttachmentMapper attachmentMapper,
-            CassandraMessageIdDAO messageIdDAO, CassandraImapUidDAO imapUidDAO) {
+    public CassandraMessageMapper(Session session, UidProvider uidProvider, ModSeqProvider modSeqProvider, MessageIdProvider messageIdProvider, MailboxSession mailboxSession, int maxRetries, AttachmentMapper attachmentMapper,
+            CassandraMessageDAO messageDAO, CassandraMessageIdDAO messageIdDAO, CassandraImapUidDAO imapUidDAO) {
         this.session = session;
         this.uidProvider = uidProvider;
         this.modSeqProvider = modSeqProvider;
         this.messageIdProvider = messageIdProvider;
         this.mailboxSession = mailboxSession;
         this.maxRetries = maxRetries;
-        this.typesProvider = typesProvider;
         this.attachmentMapper = attachmentMapper;
+        this.messageDAO = messageDAO;
         this.messageIdDAO = messageIdDAO;
         this.imapUidDAO = imapUidDAO;
     }
@@ -196,23 +171,16 @@ public class CassandraMessageMapper implements MessageMapper {
     @Override
     public Iterator<MailboxMessage> findInMailbox(Mailbox mailbox, MessageRange set, FetchType ftype, int max) throws MailboxException {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-        return CassandraUtils.convertToStream(session.execute(buildSelectQueryWithLimit(buildQuery(mailboxId, set, ftype), max)))
+        return CassandraUtils.convertToStream(messageDAO.retrieveMessages(messageIdDAO.retrieveMessageIds(mailboxId, set, ftype), ftype, Optional.of(max)))
             .map(row -> message(row, ftype))
             .sorted(Comparator.comparing(MailboxMessage::getUid))
             .iterator();
     }
 
-    private Statement buildSelectQueryWithLimit(Select.Where selectStatement, int max) {
-        if (max <= 0) {
-            return selectStatement;
-        }
-        return selectStatement.limit(max);
-    }
-
     @Override
     public List<MessageUid> findRecentMessageUidsInMailbox(Mailbox mailbox) throws MailboxException {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-        return CassandraUtils.convertToStream(session.execute(buildQuery(mailboxId, MessageRange.all(), FetchType.Metadata)))
+        return CassandraUtils.convertToStream(messageDAO.retrieveMessages(messageIdDAO.retrieveMessageIds(mailboxId, MessageRange.all(), FetchType.Metadata), FetchType.Metadata, Optional.empty()))
             .filter(row -> row.getBool(RECENT))
             .flatMap((row) -> imapUidDAO.retrieve(CassandraMessageId.of(row.getString(MESSAGE_ID)), Optional.ofNullable(mailboxId)))
             .map(UniqueMessageId::getMessageUid)
@@ -223,7 +191,7 @@ public class CassandraMessageMapper implements MessageMapper {
     @Override
     public MessageUid findFirstUnseenMessageUid(Mailbox mailbox) throws MailboxException {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-        return CassandraUtils.convertToStream(session.execute(buildQuery(mailboxId, MessageRange.all(), FetchType.Metadata)))
+        return CassandraUtils.convertToStream(messageDAO.retrieveMessages(messageIdDAO.retrieveMessageIds(mailboxId, MessageRange.all(), FetchType.Metadata), FetchType.Metadata, Optional.empty()))
             .filter(row -> !row.getBool(SEEN))
             .flatMap((row) -> imapUidDAO.retrieve(CassandraMessageId.of(row.getString(MESSAGE_ID)), Optional.ofNullable(mailboxId)))
             .map(UniqueMessageId::getMessageUid)
@@ -235,7 +203,7 @@ public class CassandraMessageMapper implements MessageMapper {
     @Override
     public Map<MessageUid, MessageMetaData> expungeMarkedForDeletionInMailbox(Mailbox mailbox, MessageRange set) throws MailboxException {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-        return CassandraUtils.convertToStream(session.execute(buildQuery(mailboxId, set, FetchType.Metadata)))
+        return CassandraUtils.convertToStream(messageDAO.retrieveMessages(messageIdDAO.retrieveMessageIds(mailboxId, set, FetchType.Metadata), FetchType.Metadata, Optional.empty()))
             .filter(row -> row.getBool(DELETED))
             .map(row -> message(row, FetchType.Metadata))
             .peek((message) -> delete(mailbox, message))
@@ -283,7 +251,7 @@ public class CassandraMessageMapper implements MessageMapper {
     @Override
     public Iterator<UpdatedFlags> updateFlags(Mailbox mailbox, FlagsUpdateCalculator flagUpdateCalculator, MessageRange set) throws MailboxException {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-        return CassandraUtils.convertToStream(session.execute(buildQuery(mailboxId, set, FetchType.Metadata)))
+        return CassandraUtils.convertToStream(messageDAO.retrieveMessages(messageIdDAO.retrieveMessageIds(mailboxId, set, FetchType.Metadata), FetchType.Metadata, Optional.empty()))
             .map((row) -> updateFlagsOnMessage(mailbox, flagUpdateCalculator, row))
             .filter(Optional::isPresent)
             .map(Optional::get)
@@ -419,61 +387,15 @@ public class CassandraMessageMapper implements MessageMapper {
     }
 
     private MessageMetaData save(Mailbox mailbox, MailboxMessage message) throws MailboxException {
-        try {
-            CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
-            insertIds(message, mailboxId);
-
-            session.execute(insertInto(TABLE_NAME)
-                .value(MESSAGE_ID, message.getMessageId().serialize())
-                .value(MOD_SEQ, message.getModSeq())
-                .value(INTERNAL_DATE, message.getInternalDate())
-                .value(BODY_START_OCTET, message.getFullContentOctets() - message.getBodyOctets())
-                .value(FULL_CONTENT_OCTETS, message.getFullContentOctets())
-                .value(BODY_OCTECTS, message.getBodyOctets())
-                .value(ANSWERED, message.isAnswered())
-                .value(DELETED, message.isDeleted())
-                .value(DRAFT, message.isDraft())
-                .value(FLAGGED, message.isFlagged())
-                .value(RECENT, message.isRecent())
-                .value(SEEN, message.isSeen())
-                .value(USER, message.createFlags().contains(Flag.USER))
-                .value(USER_FLAGS, userFlagsSet(message))
-                .value(BODY_CONTENT, toByteBuffer(message.getBodyContent()))
-                .value(HEADER_CONTENT, toByteBuffer(message.getHeaderContent()))
-                .value(PROPERTIES, message.getProperties().stream()
-                    .map(x -> typesProvider.getDefinedUserType(PROPERTIES)
-                        .newValue()
-                        .setString(Properties.NAMESPACE, x.getNamespace())
-                        .setString(Properties.NAME, x.getLocalName())
-                        .setString(Properties.VALUE, x.getValue()))
-                    .collect(Collectors.toList()))
-                .value(TEXTUAL_LINE_COUNT, message.getTextualLineCount())
-                .value(ATTACHMENTS, message.getAttachments().stream()
-                    .map(this::toUDT)
-                    .collect(Collectors.toList())));
-
-            return new SimpleMessageMetaData(message);
-        } catch (IOException e) {
-            throw new MailboxException("Error saving mail", e);
-        }
+        CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
+        insertIds(message, mailboxId);
+        messageDAO.save(mailbox, message);
+        return new SimpleMessageMetaData(message);
     }
 
     private void insertIds(MailboxMessage message, CassandraId mailboxId) {
         messageIdDAO.insert(mailboxId, message.getUid(), message.getMessageId());
         imapUidDAO.insert(message.getMessageId(), mailboxId, message.getUid());
-    }
-
-    private UDTValue toUDT(MessageAttachment messageAttachment) {
-        return typesProvider.getDefinedUserType(ATTACHMENTS)
-            .newValue()
-            .setString(Attachments.ID, messageAttachment.getAttachmentId().getId())
-            .setString(Attachments.NAME, messageAttachment.getName().orNull())
-            .setString(Attachments.CID, messageAttachment.getCid().transform(Cid::getValue).orNull())
-            .setBool(Attachments.IS_INLINE, messageAttachment.isInline());
-    }
-
-    private Set<String> userFlagsSet(MailboxMessage message) {
-        return Arrays.stream(message.createFlags().getUserFlags()).collect(Collectors.toSet());
     }
 
     private void manageUnseenMessageCounts(Mailbox mailbox, Flags oldFlags, Flags newFlags) {
@@ -499,7 +421,7 @@ public class CassandraMessageMapper implements MessageMapper {
             Flags newFlags = flagUpdateCalculator.buildNewFlags(oldFlags);
             message.setFlags(newFlags);
             message.setModSeq(modSeqProvider.nextModSeq(mailboxSession, mailbox));
-            if (conditionalSave(message, oldModSeq)) {
+            if (messageDAO.conditionalSave(message, oldModSeq)) {
                 return Optional.of(new UpdatedFlags(message.getUid(), message.getModSeq(), oldFlags, newFlags));
             } else {
                 return Optional.empty();
@@ -526,57 +448,9 @@ public class CassandraMessageMapper implements MessageMapper {
         CassandraId mailboxId = (CassandraId) mailbox.getMailboxId();
         return tryMessageFlagsUpdate(flagUpdateCalculator,
             mailbox,
-            message(Optional.ofNullable(session.execute(selectMessageData(ImmutableList.of(messageId), FetchType.Metadata)).one())
+            message(Optional.ofNullable(messageDAO.selectMessageData(ImmutableList.of(messageId), FetchType.Metadata).one())
                 .orElseThrow(() -> new MessageDeletedDuringFlagsUpdateException(mailboxId, messageId)),
                 FetchType.Metadata));
-    }
-
-    private boolean conditionalSave(MailboxMessage message, long oldModSeq) {
-        ResultSet resultSet = session.execute(
-            update(TABLE_NAME)
-                .with(set(ANSWERED, message.isAnswered()))
-                .and(set(DELETED, message.isDeleted()))
-                .and(set(DRAFT, message.isDraft()))
-                .and(set(FLAGGED, message.isFlagged()))
-                .and(set(RECENT, message.isRecent()))
-                .and(set(SEEN, message.isSeen()))
-                .and(set(USER, message.createFlags().contains(Flag.USER)))
-                .and(set(USER_FLAGS, userFlagsSet(message)))
-                .and(set(MOD_SEQ, message.getModSeq()))
-                .where(eq(MESSAGE_ID, message.getMessageId().serialize()))
-                .onlyIf(eq(MOD_SEQ, oldModSeq)));
-        return resultSet.one().getBool(CassandraConstants.LIGHTWEIGHT_TRANSACTION_APPLIED);
-    }
-
-    private ByteBuffer toByteBuffer(InputStream stream) throws IOException {
-        return ByteBuffer.wrap(ByteStreams.toByteArray(stream));
-    }
-
-    private Where buildQuery(CassandraId mailboxId, MessageRange set, FetchType fetchType) {
-        return selectMessageData(messageIdDAO.retrieveMessageIds(mailboxId, set, fetchType), fetchType);
-    }
-
-    private Where selectMessageData(List<CassandraMessageId> messageIds, FetchType fetchType) {
-        return select(retrieveFields(fetchType))
-                .from(TABLE_NAME)
-                .where(in(MESSAGE_ID, messageIds.stream()
-                        .map(CassandraMessageId::serialize)
-                        .collect(Collectors.toList())));
-    }
-
-    private String[] retrieveFields(FetchType fetchType) {
-        switch (fetchType) {
-            case Body:
-                return BODY;
-            case Full:
-                return FIELDS;
-            case Headers:
-                return HEADERS;
-            case Metadata:
-                return METADATA;
-            default:
-                throw new RuntimeException("Unknown FetchType " + fetchType);
-        }
     }
 
     private SharedByteArrayInputStream buildContent(Row row, FetchType fetchType) {
